move shared utils under ./shared
