+++ src/index.ts
@@ -1,9 +1,9 @@
 // Export all utilities
-export * from './error';
-export * from './throwIfFalsy';
-export * from './throwIfInvalid';
-export * from './throwIfNull';
-export * from './throwOnError';
-export * from './throwOnNull';
-export * from './tryCatch';
+export * from './shared/error';
+export * from './shared/throwIfFalsy';
+export * from './shared/throwIfInvalid';
+export * from './shared/throwIfNull';
+export * from './shared/throwOnError';
+export * from './shared/throwOnNull';
+export * from './shared/tryCatch';
 export * from './types';
