+++ src/internal/reportErrorShared.ts
