+++ src/shared/throwIfInvalid.ts
