+++ src/shared/throwOnNull.ts
